refactor: rename provider formats package to registry
