+++ .changeset/hook-evidence-lift-diagnostics.md
@@ -1,7 +1,7 @@
 ---
 "@skillset/core": patch
-"@skillset/provider-formats": patch
+"@skillset/registry": patch
 "skillset": patch
 ---
 
-Add structured hook provider evidence, classifier-backed native hook lift diagnostics during import/adopt, and plugin adaptive hook `run.env` rendering with deterministic shell assignments.
+Rename the built-in provider facts package to `@skillset/registry`, then add structured hook provider evidence, classifier-backed native hook lift diagnostics during import/adopt, and plugin adaptive hook `run.env` rendering with deterministic shell assignments.
