+++ apps/skillset/package.json
@@ -35,7 +35,7 @@
   "devDependencies": {
     "@skillset/core": "workspace:*",
     "@skillset/lint": "workspace:*",
-    "@skillset/provider-formats": "workspace:*",
+    "@skillset/registry": "workspace:*",
     "@skillset/schema": "workspace:*",
     "@skillset/toolkit": "workspace:*",
     "@skillset/transforms": "workspace:*"
