+++ apps/skillset/src/changeset-awareness.ts
@@ -48,7 +48,7 @@ export function isPackageAffectingPath(path: string) {
   if (isRuntimeSourcePath(path, "apps/skillset/src")) return true;
   if (isRuntimeSourcePath(path, "packages/core/src")) return true;
   if (isRuntimeSourcePath(path, "packages/lint/src")) return true;
-  if (isRuntimeSourcePath(path, "packages/provider-formats/src")) return true;
+  if (isRuntimeSourcePath(path, "packages/registry/src")) return true;
   if (isRuntimeSourcePath(path, "packages/schema/src")) return true;
   if (isRuntimeSourcePath(path, "packages/toolkit/src")) return true;
   if (isRuntimeSourcePath(path, "packages/transforms/src")) return true;
@@ -56,7 +56,7 @@ export function isPackageAffectingPath(path: string) {
   return (
     path === "packages/core/package.json" ||
     path === "packages/lint/package.json" ||
-    path === "packages/provider-formats/package.json" ||
+    path === "packages/registry/package.json" ||
     path === "packages/schema/package.json" ||
     path === "packages/toolkit/package.json" ||
     path === "packages/transforms/package.json"
