+++ docs/package-releases.md
@@ -13,7 +13,7 @@ GitHub Actions is the package release operator. Local commands are diagnostics a
 
 Changesets owns npm package version and package changelog calculation. The Skillset change/release commands continue to own source-unit reasons, release state, generated entity changelogs, and target output drift. Do not collapse these two release systems unless a future explicit bridge is designed.
 
-Only the unscoped `skillset` CLI package is public in the current package posture. The workspace packages `@skillset/core`, `@skillset/lint`, and `@skillset/transforms` remain private implementation packages: core is the internal library boundary that the CLI and tests consume, while lint and transforms are support packages behind that boundary. Do not include them in npm publish automation or treat their exports as semver-stable until a future package-posture issue explicitly promotes them.
+Only the unscoped `skillset` CLI package is public in the current package posture. The scoped workspace packages remain private implementation packages: `@skillset/core` is the internal compiler/library boundary, `@skillset/registry` stores deterministic provider facts and registry contracts, and the remaining scoped packages support schema, lint, toolkit, transform, and workbench surfaces behind that boundary. Do not include them in npm publish automation or treat their exports as semver-stable until a future package-posture issue explicitly promotes them.
 
 ## Flow
 
@@ -27,11 +27,11 @@ Package-facing means a change that can affect the published `skillset` CLI packa
 | `apps/skillset/package.json` | Published package metadata, bin entries, dependencies, and version-bearing state. |
 | `packages/core/src/**` except tests | Internal compiler/library implementation bundled through the CLI. |
 | `packages/lint/src/**` except tests | Lint implementation consumed by the CLI. |
-| `packages/provider-formats/src/**` except tests | Adopted provider destination-format snapshots, schema evidence, and migration registry consumed by the CLI and core conformance checks. |
+| `packages/registry/src/**` except tests | Adopted provider destination-format snapshots, schema evidence, and migration registry consumed by the CLI and core conformance checks. |
 | `packages/schema/src/**` except tests | Source contract schemas, validators, examples, and artifact generation consumed by the CLI, Workbench, and generated editor-schema references. |
 | `packages/toolkit/src/**` except tests | Runtime helper surfaces intended for hook scripts and compiler-owned wrappers. |
 | `packages/transforms/src/**` except tests | Transform implementation consumed by the CLI. |
-| `packages/*/package.json` for `core`, `lint`, `provider-formats`, `schema`, and `transforms` | Runtime dependency and package metadata for the private workspace packages that feed the CLI. |
+| `packages/*/package.json` for `core`, `lint`, `registry`, `schema`, and `transforms` | Runtime dependency and package metadata for the private workspace packages that feed the CLI. |
 | `bun.lock` / `bun.lockb` | Dependency resolution that can alter the packaged CLI runtime. |
 
 `bun run changeset:check` enforces this boundary. It fails when package-facing paths change without an active `.changeset/*.md`, and it also fails when an active Changeset appears on a branch that only changes repo machinery. Deleted Changesets are ignored so cleanup branches can remove mistaken package-release entries.
@@ -91,11 +91,11 @@ bun run publish:registry-check:published
 
 `bun run publish:check` is the local preflight: it runs the full repo check, rebuilds the npm package output, and performs `bun pm pack --dry-run` from `apps/skillset` so package contents are verified without registry authentication.
 
-Before marking a release PR ready, review provider and schema evidence when the range touches `packages/provider-formats/src/**`, `packages/schema/src/**`, `docs/reference/schemas/**`, or `docs/reference/examples/**`:
+Before marking a release PR ready, review provider and schema evidence when the range touches `packages/registry/src/**`, `packages/schema/src/**`, `docs/reference/schemas/**`, or `docs/reference/examples/**`:
 
 1. Run `bun run schema:check` for schema contract or generated schema/example changes.
 2. Run `bun ./apps/skillset/src/cli.ts providers check --root .` and, when upstream drift is expected, `bun ./apps/skillset/src/cli.ts providers diff --root .` for provider snapshot or migration changes.
-3. Inspect `packages/provider-formats/src/{index.ts,schema-snapshots.ts,migrations.ts}` alongside `docs/target-surfaces.md` so the adopted snapshot, migration class, and user-facing diagnostics tell the same story.
+3. Inspect `packages/registry/src/{index.ts,schema-snapshots.ts,migrations.ts}` alongside `docs/target-surfaces.md` so the adopted snapshot, migration class, and user-facing diagnostics tell the same story.
 4. Confirm each package-facing provider/schema change has a `.changeset/*.md` entry using the wording class above, and each local source-unit behavior change has a Skillset pending change entry where appropriate.
 5. Keep generated `docs/reference/schemas/**` and `docs/reference/examples/**` diffs in the same branch as their `packages/schema/src/**` source change, then rerun `bun run changeset:check` and `bun run skillset:verify`.
