+++ docs/target-surfaces.md
@@ -1,18 +1,18 @@
 # Provider Surface Evidence Matrix
 
-This is the cheap-to-refresh map between Skillset source and the Claude/Codex provider surfaces it builds. It exists so provider drift is caught deliberately: each surface row has a **status** and, where it depends on provider-owned destination formats, an adopted snapshot in `@skillset/provider-formats`. Golden manifest tests in `apps/skillset/src/__tests__/contract.test.ts` and `apps/skillset/src/__tests__/skillset.test.ts` pin the generated shapes that these rows claim.
+This is the cheap-to-refresh map between Skillset source and the Claude/Codex provider surfaces it builds. It exists so provider drift is caught deliberately: each surface row has a **status** and, where it depends on provider-owned destination formats, an adopted snapshot in `@skillset/registry`. Golden manifest tests in `apps/skillset/src/__tests__/contract.test.ts` and `apps/skillset/src/__tests__/skillset.test.ts` pin the generated shapes that these rows claim.
 
 Refreshing is intentionally cheap but explicit: re-read the linked provider docs, update the normalized snapshot with its fetched timestamp and content hash, then adjust the affected row and golden test if the surface changed. Ordinary build and check paths must not fetch provider docs.
 
-`@skillset/provider-formats` also records rolling-latest provider schema sources for maintainer checks: Claude Code settings, plugin manifests, marketplaces, and keybindings; Codex config, hooks config, hook event schemas, and skill metadata. Those schema snapshots store source URLs, upstream content hashes, fetched timestamps, normalized summaries, and deterministic snapshot hashes. Destination areas with no adopted JSON Schema source are represented as manual overlays linked back to the docs-backed format snapshots rather than left as silent gaps.
+`@skillset/registry` also records rolling-latest provider schema sources for maintainer checks: Claude Code settings, plugin manifests, marketplaces, and keybindings; Codex config, hooks config, hook event schemas, and skill metadata. Those schema snapshots store source URLs, upstream content hashes, fetched timestamps, normalized summaries, and deterministic snapshot hashes. Destination areas with no adopted JSON Schema source are represented as manual overlays linked back to the docs-backed format snapshots rather than left as silent gaps.
 
-Maintainers refresh that evidence explicitly with `skillset providers check`, `skillset providers diff`, and `skillset providers update --yes`. `check` compares adopted schema source hashes against live upstream sources, `diff` adds readable schema summary changes plus explicit manual-review rows for prose-only destination format snapshots, and `update --yes` rewrites `packages/provider-formats/src/schema-snapshots.ts` with refreshed schema summaries and provenance. Destination format snapshots are not machine-diffed until their source rows carry a machine-readable upstream baseline. Ordinary user builds, checks, and generated-output verification remain offline.
+Maintainers refresh that evidence explicitly with `skillset providers check`, `skillset providers diff`, and `skillset providers update --yes`. `check` compares adopted schema source hashes against live upstream sources, `diff` adds readable schema summary changes plus explicit manual-review rows for prose-only destination format snapshots, and `update --yes` rewrites `packages/registry/src/schema-snapshots.ts` with refreshed schema summaries and provenance. Destination format snapshots are not machine-diffed until their source rows carry a machine-readable upstream baseline. Ordinary user builds, checks, and generated-output verification remain offline.
 
-Known safe update paths live in `packages/provider-formats/src/migrations.ts`. Add a registry entry when a provider destination-format change is understood well enough to classify as `compatible`, `adapter-only`, `source-migration`, `unsupported-drift`, or `manual-review`. Each entry must name the provider, destination surface, source and destination snapshot versions, safety and preview flags, confirmation requirement, and the adopted snapshot ids it applies to. Leave unreviewed or lossy changes unregistered so callers route them to manual review instead of rewriting source.
+Known safe update paths live in `packages/registry/src/migrations.ts`. Add a registry entry when a provider destination-format change is understood well enough to classify as `compatible`, `adapter-only`, `source-migration`, `unsupported-drift`, or `manual-review`. Each entry must name the provider, destination surface, source and destination snapshot versions, safety and preview flags, confirmation requirement, and the adopted snapshot ids it applies to. Leave unreviewed or lossy changes unregistered so callers route them to manual review instead of rewriting source.
 
 `skillset check` reports generated-output drift that maps to safe provider destination-format migrations without writing. `skillset check --fix` and `skillset update --yes` apply only source-preserving safe plans; unregistered or manual-review drift blocks writes and prints the affected outputs.
 
-Provider-format evidence and migration changes are package-facing when they touch `packages/provider-formats/src/**` or `packages/provider-formats/package.json`. Record compatible refreshes, safe destination-format migrations, and manual-review drift in the package Changeset with the same class names that users see in diagnostics. Add a Skillset pending change entry when the target rendering model, provider support promise, or generated-output behavior changes in this workspace; leave pure docs clarification as docs-only unless it changes release-visible behavior.
+Provider-format evidence and migration changes are package-facing when they touch `packages/registry/src/**` or `packages/registry/package.json`. Record compatible refreshes, safe destination-format migrations, and manual-review drift in the package Changeset with the same class names that users see in diagnostics. Add a Skillset pending change entry when the target rendering model, provider support promise, or generated-output behavior changes in this workspace; leave pure docs clarification as docs-only unless it changes release-visible behavior.
 
 User-facing destination-format diagnostics should name the provider, destination, source unit, affected output, and next command or review step. Safe preview output looks like:
